+++ source/main/index.js
@@ -22,12 +22,15 @@ import {
   launcherConfig,
   pubLogsFolderPath,
   stateDirectoryPath,
+  RTS_FLAGS,
+  MINIMUM_AMOUNT_OF_RAM_FOR_RTS_FLAGS,
 } from './config';
 import { setupCardanoNode } from './cardano/setup';
 import { CardanoNode } from './cardano/CardanoNode';
 import { safeExitWithCode } from './utils/safeExitWithCode';
 import { buildAppMenus } from './utils/buildAppMenus';
 import { getLocale } from './utils/getLocale';
+import { getRtsFlags, setRtsFlagsAndRestart } from './utils/rtsFlags';
 import { detectSystemLocale } from './utils/detectSystemLocale';
 import { ensureXDGDataIsSet } from './cardano/config';
 import { rebuildApplicationMenu } from './ipc/rebuild-application-menu';
@@ -117,6 +120,7 @@ const onAppReady = async () => {
   const cpu = os.cpus();
   const platformVersion = os.release();
   const ram = JSON.stringify(os.totalmem(), null, 2);
+
   const startTime = new Date().toISOString();
   // first checks for Japanese locale, otherwise returns english
   const systemLocale = detectSystemLocale();
@@ -162,8 +166,23 @@ const onAppReady = async () => {
   );
   saveWindowBoundsOnSizeAndPositionChange(mainWindow, requestElectronStore);
 
-  logger.info('Setting up Cardano Node...');
-  cardanoNode = setupCardanoNode(launcherConfig, mainWindow);
+  const getCurrentRtsFlags = () => {
+    const rtsFlagsFromStorage = getRtsFlags(network);
+    if (!rtsFlagsFromStorage) {
+      if (os.totalmem() < MINIMUM_AMOUNT_OF_RAM_FOR_RTS_FLAGS) {
+        setRtsFlagsAndRestart(environment.network, RTS_FLAGS);
+        return RTS_FLAGS;
+      }
+      return [];
+    }
+    return rtsFlagsFromStorage;
+  };
+
+  const rtsFlags = getCurrentRtsFlags();
+  logger.info(
+    `Setting up Cardano Node... with flags: ${JSON.stringify(rtsFlags)}`
+  );
+  cardanoNode = setupCardanoNode(launcherConfig, mainWindow, rtsFlags);
 
   buildAppMenus(mainWindow, cardanoNode, userLocale, {
     isNavigationEnabled: false,